refactoring toolbar customize
update toolbar ui
fix some problems
