+++ source/MRViewer/MRViewer.vcxproj
@@ -23,6 +23,7 @@
     <ClCompile Include="MRMouseController.cpp" />
     <ClCompile Include="MRRenderVolumeObject.cpp" />
     <ClCompile Include="MRSceneSelectionChange.cpp" />
+    <ClCompile Include="MRToolbar.cpp" />
     <ClCompile Include="MRTouchesHandler.cpp" />
     <ClCompile Include="MRTouchesController.cpp" />
     <ClCompile Include="MRObjectMeshHistory.cpp" />
@@ -91,6 +92,7 @@
     <ClInclude Include="MRMouseController.h" />
     <ClInclude Include="MRRenderVolumeObject.h" />
     <ClInclude Include="MRSceneSelectionChange.h" />
+    <ClInclude Include="MRToolbar.h" />
     <ClInclude Include="MRTouchesController.h" />
     <ClInclude Include="MRObjectMeshHistory.h" />
     <ClInclude Include="MRObjectTransformWidget.h" />
